Update index.html.md for security-alerts-api
Update the index.html.md file for security-alerts-api with corrected section names for "Filter syntax for the query-alerts API command" and "Filter syntax for the rules/query-rule-metadata API command". Those sections contain hyphens and a slash which mean that the sections don't appear in the left navigation, nor do links to them work. The update removes the special characters from the section names.